+++ crates/cxx-qt-gen/src/generator/rust/qobject.rs
@@ -9,7 +9,7 @@ use crate::{
         rust::{
             field::generate_rust_fields, fragment::RustFragmentPair, inherit,
             invokable::generate_rust_invokables, property::generate_rust_properties,
-            signals::generate_rust_signals,
+            signals::generate_rust_signals, threading,
         },
     },
     parser::qobject::ParsedQObject,
@@ -36,10 +36,6 @@ impl GeneratedRustQObjectBlocks {
 pub struct GeneratedRustQObject {
     /// Ident of the Rust name for the C++ object
     pub cpp_struct_ident: Ident,
-    /// Ident of the CxxQtThread object
-    pub cxx_qt_thread_ident: Ident,
-    /// Ident of the Rust closure wrapper to be passed in to CxxQtThread
-    pub cxx_qt_thread_queued_fn_ident: Ident,
     /// Ident of the namespace for CXX-Qt internals of the QObject
     pub namespace_internals: String,
     /// Ident of the Rust name for the Rust object
@@ -55,9 +51,7 @@ impl GeneratedRustQObject {
         let namespace_idents = NamespaceName::from(qobject);
         let mut generated = GeneratedRustQObject {
             cpp_struct_ident: qobject_idents.cpp_class.rust.clone(),
-            cxx_qt_thread_ident: qobject_idents.cxx_qt_thread_class.clone(),
-            cxx_qt_thread_queued_fn_ident: qobject_idents.cxx_qt_thread_queued_fn_struct.clone(),
-            namespace_internals: namespace_idents.internal,
+            namespace_internals: namespace_idents.internal.clone(),
             rust_struct_ident: qobject_idents.rust_struct.rust.clone(),
             blocks: GeneratedRustQObjectBlocks {
                 cxx_mod_contents: vec![],
@@ -118,6 +112,14 @@ impl GeneratedRustQObject {
                     .cxx_mod_contents
                     .append(&mut fragment.cxx_bridge_as_items()?);
             }
+        }
+
+        // If this type has threading enabled then add generation
+        if qobject.threading {
+            generated.blocks.append(&mut threading::generate(
+                &qobject_idents,
+                &namespace_idents,
+            )?);
         }
 
         Ok(generated)
@@ -211,11 +213,6 @@ mod tests {
         let rust = GeneratedRustQObject::from(parser.cxx_qt_data.qobjects.values().next().unwrap())
             .unwrap();
         assert_eq!(rust.cpp_struct_ident, "MyObjectQt");
-        assert_eq!(rust.cxx_qt_thread_ident, "MyObjectCxxQtThread");
-        assert_eq!(
-            rust.cxx_qt_thread_queued_fn_ident,
-            "MyObjectCxxQtThreadQueuedFn"
-        );
         assert_eq!(rust.namespace_internals, "cxx_qt_my_object");
         assert_eq!(rust.rust_struct_ident, "MyObject");
     }
@@ -234,11 +231,6 @@ mod tests {
         let rust = GeneratedRustQObject::from(parser.cxx_qt_data.qobjects.values().next().unwrap())
             .unwrap();
         assert_eq!(rust.cpp_struct_ident, "MyObjectQt");
-        assert_eq!(rust.cxx_qt_thread_ident, "MyObjectCxxQtThread");
-        assert_eq!(
-            rust.cxx_qt_thread_queued_fn_ident,
-            "MyObjectCxxQtThreadQueuedFn"
-        );
         assert_eq!(rust.namespace_internals, "cxx_qt::cxx_qt_my_object");
         assert_eq!(rust.rust_struct_ident, "MyObject");
     }