+++ crates/cxx-qt-gen/test_outputs/signals.h
@@ -10,7 +10,7 @@ class CxxQtThread;
 
 namespace cxx_qt::my_object {
 class MyObject;
-using MyObjectCxxQtThread = ::rust::cxxqtlib1::CxxQtThread<MyObject>;
+
 } // namespace cxx_qt::my_object
 
 #include "cxx-qt-gen/ffi.cxx.h"
@@ -25,7 +25,6 @@ public:
   ~MyObject();
   MyObjectRust const& unsafeRust() const;
   MyObjectRust& unsafeRustMut();
-  ::std::unique_ptr<MyObjectCxxQtThread> qtThread() const;
 
 public:
   Q_INVOKABLE void invokable();
@@ -63,8 +62,6 @@ public:
 private:
   ::rust::Box<MyObjectRust> m_rustObj;
   ::std::shared_ptr<::std::recursive_mutex> m_rustObjMutex;
-  ::std::shared_ptr<::rust::cxxqtlib1::CxxQtGuardedPointer<MyObject>>
-    m_cxxQtThreadObj;
 };
 
 static_assert(::std::is_base_of<QObject, MyObject>::value,